refactor: move Home top menu
move Home top menu into separate composable
